+++ ci/update-derived-files.sh
@@ -23,7 +23,7 @@ for cmd in start ssh serial connect ls-fonts show-keys imgcat set-working-direct
   cargo run --example narrow $PWD/target/debug/wezterm $cmd --help | ./target/debug/strip-ansi-escapes > $fname
 done
 
-for cmd in list list-clients move-pane-to-new-tab split-pane spawn send-text get-text activate-pane-direction kill-pane activate-pane activate-tab ; do
+for cmd in list list-clients move-pane-to-new-tab split-pane spawn send-text get-text activate-pane-direction kill-pane activate-pane activate-tab set-tab-title set-window-title ; do
   fname="docs/examples/cmd-synopsis-wezterm-cli-${cmd}--help.txt"
   cargo run --example narrow $PWD/target/debug/wezterm cli $cmd --help | ./target/debug/strip-ansi-escapes > $fname
 done